requirements.txt
-i https://pypi.org/simple
backcall==0.1.0
blessings==1.7
click==7.0
coverage==4.5.1
decorator==4.4.0
dj-database-url==0.5.0
django-configurations==2.0
django-filter==2.0.0
django-model-utils==3.1.2
django-nose==1.4.5
django-storages==1.6.6
django-unique-upload==0.2.1
django==2.0.7
djangorestframework==3.8.2
factory-boy==2.11.1
faker==2.0.1
flake8==3.5.0
gunicorn==19.8.1
ipdb==0.11
ipython-genutils==0.2.0
ipython==6.5.0
jedi==0.15.1
jinja2==2.10.1
livereload==2.6.1
markdown==2.6.11
markupsafe==1.1.1
mccabe==0.6.1
mkdocs==0.17.5
mock==2.0.0
newrelic==4.0.0.99
nose-progressive==1.5.1
nose==1.3.7
parso==0.5.1
pbr==5.4.2
pexpect==4.7.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==1.0.16
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
pygments==2.4.2
python-dateutil==2.8.0
pytz==2018.5
pyyaml==5.1.2
simplegeneric==0.8.1
six==1.12.0
text-unidecode==1.2
tornado==4.5.3
traitlets==4.3.2
wcwidth==0.1.7
